bot/requirements.txt
maxapi>=0.9.16
sqlalchemy>=2.0
gspread>=6.0
google-auth>=2.0
